Split classes and fix getInput function
