+++ composer.json
@@ -5,9 +5,6 @@
     "license": "GPL-3.0",
     "minimum-stability": "dev",
     "prefer-stable": true,
-    "require": {
-        "guzzlehttp/guzzle": "^6.3"
-    },
     "require-dev": {
         "phpstan/phpstan": "^0.10.7",
         "phpunit/phpunit": "^7",
