+++ src/Model/Element.php
@@ -1,0 +1,30 @@
+<?php
+
+/*****************************************************************************
+ *
+ *  PROJECT:     MTASA PHP SDK
+ *  LICENSE:     See LICENSE in the top level directory
+ *  FILE:        Element.php
+ *
+ *  Multi Theft Auto is available from http://www.multitheftauto.com/
+ *
+ *****************************************************************************/
+
+declare(strict_types=1);
+
+namespace MultiTheftAuto\Sdk\Model;
+
+class Element
+{
+    var $id;
+
+    function __construct($id)
+    {
+        $this->id = $id;
+    }
+
+    function toString()
+    {
+        return "^E^" . $this->id;
+    }
+}
